refreshed apis

+++ src/main/java/com/secure/secure/controller/GroupController.java
@@ -3,12 +3,17 @@ package com.secure.secure.controller;
 import com.secure.secure.dto.Input;
 import com.secure.secure.dto.Output;
 import com.secure.secure.entity.Group;
+import com.secure.secure.entity.GroupJoinRequest;
+import com.secure.secure.repository.GroupJoinRequestRepository;
 import com.secure.secure.repository.GroupRepository;
+import com.secure.secure.repository.PostRepository;
 import com.secure.secure.repository.UserRepository;
 import com.secure.secure.util.CustomException;
 import com.secure.secure.util.ResponseMapper;
 import com.secure.secure.util.ResponseObject;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Date;
@@ -23,18 +28,39 @@ public class GroupController implements ResponseMapper {
     @Autowired
     private GroupRepository groupRepository;
 
+    @Autowired
+    private PostRepository postRepository;
+    @Autowired
+    private GroupJoinRequestRepository groupJoinRequestRepository;
+
+    public String getUsername(){
+        Object principal = SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        String username ="";
+        if (principal instanceof UserDetails) {
+            username = ((UserDetails)principal).getUsername();
+        } else {
+            username = principal.toString();
+        }
+        return username;
+    }
+
     @PostMapping("/create")
     public ResponseObject createGroup(@RequestBody Input.CreateGroup input){
-        //todo validate each input and add limit to each input
-        //todo check if group exist with same name
-        Group group = new Group();
-        group.setGroupName(input.groupName());
-        group.setGroupDescription(input.groupDescription());
-        group.setCreatedBy(input.createdBy());
-        group.setCreatedTime(new Date());
-        var user = userRepository.findByUsername(input.createdBy());
-        group.setGroupMembers(List.of(user.get()));
-        return successResponse("Created Group: "+ input.groupName());
+        try{
+            var user = userRepository.findByUsername(getUsername());
+            Group group = new Group();
+            group.setGroupName(input.groupName());
+            group.setGroupDescription(input.groupDescription());
+            group.setCreatedBy(user.get().getUsername());
+            group.setCreatedTime(new Date());
+            group.setGroupMembers(List.of(user.get()));
+            var saved = groupRepository.save(group);
+            return successResponse("Created Group: "+ saved.getGroupName());
+        }
+        catch (Exception e ){
+            return errorResponse(new CustomException("Server Error..."));
+        }
     }
 
     @PutMapping("/update-description")
@@ -49,17 +75,18 @@ public class GroupController implements ResponseMapper {
         groupRepository.save(group.get());
         return successResponse("Updated Description: "+ group.get().getGroupName());
     }
-
+/*
     @PutMapping("/add-user")
     public ResponseObject addUser(@RequestBody Input.UserGroup input){
-        //todo validate each input and add limit to each input
-        //todo only create should be able to edit
         var group = groupRepository.findById(input.groupId());
-        if(!group.isPresent()){
-            return errorResponse(new CustomException("Group not Found"));
+        var user = userRepository.findByUsername(getUsername());
+        if(!user.get().getUsername().equals("defaultAdmin") ||
+                !group.get().getCreatedBy().equals(getUsername())){
+            return errorResponse(new CustomException("Not Authorized"));
         }
-        var user = userRepository.findById(input.userId());
         group.get().getGroupMembers().add(user.get());
+        //delete from group request`
+
         groupRepository.save(group.get());
         return successResponse("Added User to group");
     }
@@ -76,7 +103,7 @@ public class GroupController implements ResponseMapper {
         group.get().getGroupMembers().remove(user.get());
         groupRepository.save(group.get());
         return successResponse("Removed User from group");
-    }
+    }*/
 
     @PutMapping("/update-limit")
     public ResponseObject updateLimit(@RequestBody Input.UserGroup input){
@@ -93,8 +120,14 @@ public class GroupController implements ResponseMapper {
 
     @DeleteMapping("/delete/{id}")
     public ResponseObject deleteGroup(@PathVariable(name ="id") String groupId){
-        //todo validate userId
-        //todo remove all group posts
+        var group = groupRepository.findById(groupId);
+        var user = userRepository.findByUsername(getUsername());
+        if(!user.get().getUsername().equals("defaultAdmin") ||
+                !group.get().getCreatedBy().equals(getUsername())){
+            return errorResponse(new CustomException("Not Authorized"));
+        }
+        groupJoinRequestRepository.deleteAll(groupJoinRequestRepository.getRequest(groupId));
+        postRepository.deleteAll(postRepository.getAllGroupPosts(groupId));
         groupRepository.deleteById(groupId);
         return successResponse("Successfully deleted the group");
 
@@ -102,24 +135,43 @@ public class GroupController implements ResponseMapper {
 
     @GetMapping("/all")
     public ResponseObject allGroups(){
-        //todo validate admin
+        var user = userRepository.findByUsername(getUsername());
+        if(!user.get().getUsername().equals("defaultAdmin")){
+            return errorResponse(new CustomException("Not Authorized"));
+        }
         return successResponse(groupRepository.findAll().stream()
                 .map(x-> new Output.GroupsList(x.getId(),x.getGroupName(),x.getGroupDescription())));
     }
 
-    @GetMapping("/user-all/{id}")
-    public ResponseObject allUserGroups(@PathVariable(name ="id") String userId){
-        //todo validate admin
-        return successResponse(groupRepository.findUserGroups(userId).stream()
+    @GetMapping("/currentUser")
+    public ResponseObject allUserGroups(){
+        return successResponse(groupRepository.findUserGroups1(getUsername()).stream()
                 .map(x-> new Output.GroupsList(x.getId(),x.getGroupName(),x.getGroupDescription())));
     }
 
-    @GetMapping("/explore/{id}")
-    public ResponseObject exploreGroups(@PathVariable(name ="id") String userId){
+    @GetMapping("/explore")
+    public ResponseObject exploreGroups(){
         //todo validate admin
-        var userGroups = groupRepository.findUserGroups(userId).stream().map(Group::getId).toList();
+        var userGroups = groupRepository.findUserGroups1(getUsername()).stream().map(Group::getId).toList();
         return successResponse(groupRepository.findAll().stream()
                 .filter(x->!userGroups.contains(x.getId()))
                 .map(x-> new Output.GroupsList(x.getId(),x.getGroupName(),x.getGroupDescription())));
     }
+
+    @PostMapping("/request/{id}")
+    public ResponseObject requestGroup(@PathVariable(name ="id") String groupId){
+        try {
+            var user = userRepository.findByUsername(getUsername());
+            GroupJoinRequest groupJoinRequest = new GroupJoinRequest();
+            groupJoinRequest.setGroup(groupRepository.findById(groupId).get());
+            groupJoinRequest.setUser(user.get());
+            groupJoinRequestRepository.save(groupJoinRequest);
+            return successResponse("Request Submitted");
+        }
+        catch (Exception e){
+            return errorResponse(new CustomException("Server Error!"));
+        }
+    }
+
+
 }
